break up config tests
